+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * CrawlerController->getLogEntriesForPageId()
 * QueueRepository->countAllByProcessId()
 * QueueRepository->countUnprocessedItems()
+* FlushQueueCommand --page option
 
 ### Fixed
 * Typecasting port number in cli/bootstrap to have processQueue working with non-standard ports